pyproject.toml
[tool.poetry]
name = "lanco-bot"
version = "0.1.0"
description = ""
authors = ["Nate Shoffner <nate.shoffner@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.5.2"
lcwc = "0.12.3"
pydantic = "^2.4.2"
peewee = "^3.17.6"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
pytz = "^2023.3.post1"
toml = "^0.10.2"
googlemaps = "^4.10.0"
confusables = "^1.2.0"
nest-asyncio = "^1.5.8"
asyncpraw = "*"
zalgo-text = "^0.6"
cachetools = "^5.3.2"
pyowm = "^3.3.0"
opencage = "^2.4.0"
spotipy = "^2.25.1"
watchfiles = "^0.21.0"
pymupdf = "^1.24.2"
vt-py = "^0.18.2"
bs4 = "^0.0.2"
emoji = "^2.13.2"
imageio = "^2.35.1"
pillow-heif = "^0.18.0"
logtail-python = "^0.3.1"
pydub = "^0.25.1"
seeclickfix = "0.3"
reactionmenu = "^3.1.7"
jikanpy-v4 = "^1.0.2"
aiogoogle = "^5.13.2"
pymysql = "^1.1.1"
cairosvg = "^2.7.1"
num2words = "^0.5.14"
wolframalpha = "^5.1.3"
elasticsearch = "^9.0.2"
pydantic-ai = "^0.3.2"
psutil = "^7.0.0"
everbridge = "^0.2.0"
sightengine-async = "^0.2"
truthbrush = { git = "https://github.com/NateShoffner/truthbrush.git" }
beautifulsoup4 = "^4.13.4"
tabulate = "^0.9.0"
setuptools = "^80.9.0"
[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^24.3.0"
isort = "^5.13.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
migrate-db = "migrate:run_migrations"
bot = "app.main:main"
cog = "tools.cog:main"